Moved billing to its own folder and adjusted path
