+++ client/src/components/Billing/billing.js
@@ -3,11 +3,11 @@ import { injectStripe } from "react-stripe-elements";
 import axios from "axios";
 import CheckoutForm from "./checkoutForm";
 
-import Sidebar from "./subComponents/sidebar";
-import Navbar from "./subComponents/navbar";
-import Loading from "./subComponents/Loading"
-import "./CSS/billing.css";
-const urls = require("../config/config.json");
+import Sidebar from "../subComponents/sidebar";
+import Navbar from "../subComponents/navbar";
+import Loading from "./Loading"
+import "./billing.css";
+const urls = require("../../config/config.json")
 
 class Billing extends Component {
   state = {
